Add files via upload
This software prompts the user to enter total number of published centers, total number of registered voters, total number of null and void votes, total number of valid votes and total valid votes for each candidate. Afterward, it determines the candidate with the majority of votes and displays the results on the screen.